V2 changes, v1 to install and launch python packages

+++ build-and-install/src/main/phases/phases.ts
@@ -33,6 +33,7 @@ import {
 	PackageType_FirebaseFunctionsApp,
 	PackageType_FirebaseHostingApp,
 	PackageType_InfraLib,
+	PackageType_Python,
 	RuntimePackage_WithOutput
 } from '../core/types';
 import {
@@ -55,6 +56,7 @@ import {MemKey_ProjectScreen} from '../screen/ProjectScreen';
 import {Commando} from '@nu-art/commando/core/cli';
 import {RunningProcessLogs} from '../screen/RunningProcessLogs';
 import {CommandExecutor_FirebaseFunction, CommandExecutor_FirebaseHosting} from '../logic/command-executors';
+import {CommandExecutor_Python} from '../logic/command-executors/CommandExecutor_Python';
 
 
 const CONST_ThunderstormVersionKey = 'THUNDERSTORM_SDK_VERSION';
@@ -141,6 +143,7 @@ export const Phase_PrepareParams: BuildPhase = {
 	isMandatory: true,
 	breakAfterPhase: true,
 	mandatoryPhases: [Phase_SetupProject, Phase_SetWithThunderstorm],
+	filter: async (pkg) => pkg.type !== PackageType_Python,
 	action: async (pkg) => {
 		const packages = MemKey_Packages.get();
 		const projectScreen = MemKey_ProjectScreen.get();
@@ -165,6 +168,9 @@ export const Phase_ResolveTemplate: BuildPhase = {
 	isMandatory: true,
 	mandatoryPhases: [Phase_PrepareParams, Phase_SetupProject, Phase_SetWithThunderstorm],
 	action: async (pkg) => {
+		if (pkg.type === PackageType_Python)
+			return;
+
 		const packages = MemKey_Packages.get();
 		const projectScreen = MemKey_ProjectScreen.get();
 
@@ -297,6 +303,9 @@ export const Phase_CheckCyclicImports: BuildPhase = {
 	mandatoryPhases: [Phase_ResolveEnv],
 	filter: async (pkg) => RuntimeParams.checkCyclicImports,
 	action: async (pkg) => {
+		if (pkg.type === PackageType_Python)
+			return;
+
 		if (!pkg.output)
 			return;
 
@@ -327,8 +336,10 @@ export const Phase_PackagePurge: BuildPhase = {
 	type: PackageBuildPhaseType_PackageWithOutput,
 	name: 'package-purge',
 	mandatoryPhases: [Phase_ResolveEnv],
-	filter: async (pkg) => fs.existsSync(pkg.output) && RuntimeParams.purge,
+	filter: async (pkg) => pkg.type !== PackageType_Python && fs.existsSync(pkg.output) && RuntimeParams.purge,
 	action: async (pkg) => {
+		if (pkg.type === PackageType_Python)
+			return;
 		const projectScreen = MemKey_ProjectScreen.get();
 
 		//Update cli ui
@@ -375,12 +386,33 @@ export const Phase_InstallPackages: BuildPhase = {
 	}
 };
 
+export const Phase_InstallPythonPackages: BuildPhase = {
+	type: 'package',
+	name: 'install-python-packages',
+	filter: async pkg => RuntimeParams.encounterManager && pkg.type === PackageType_Python,
+	action: async (pkg) => {
+		if (pkg.type !== PackageType_Python)
+			return;
+
+		await NVM.createCommando(Cli_Basic)
+			.cd(convertToFullPath(pkg.path))
+			.append('echo installing python')
+			.append('python3 -m venv venv')
+			.append('source venv/bin/activate')
+			.append('pip3 install -r requirements.txt')
+			.execute();
+	}
+};
+
 export const Phase_Clean: BuildPhase = {
 	type: PackageBuildPhaseType_PackageWithOutput,
 	name: 'clean',
 	mandatoryPhases: [Phase_ResolveEnv],
 	filter: async (pkg) => RuntimeParams.clean,
 	action: async (pkg) => {
+		if (pkg.type === PackageType_Python)
+			return;
+
 		const projectScreen = MemKey_ProjectScreen.get();
 
 		projectScreen.updateOrCreatePackage(pkg.name, 'Cleaning');
@@ -397,6 +429,9 @@ export const Phase_Lint: BuildPhase = {
 	mandatoryPhases: [Phase_ResolveEnv],
 	filter: async (pkg) => RuntimeParams.lint && pkg.type !== 'sourceless',
 	action: async (pkg) => {
+		if (pkg.type === PackageType_Python)
+			return;
+
 		const projectScreen = MemKey_ProjectScreen.get();
 
 		projectScreen.updateOrCreatePackage(pkg.name, 'Linting');
@@ -432,9 +467,9 @@ export const Phase_PrepareCompile: BuildPhase = {
 	name: 'prepare-compile',
 	isMandatory: true,
 	mandatoryPhases: [Phase_ResolveEnv],
-	filter: async (pkg) => pkg.type !== 'sourceless' && !RuntimeParams.noBuild,
+	filter: async (pkg) => pkg.type !== 'sourceless' && pkg.type !== PackageType_Python && !RuntimeParams.noBuild,
 	action: async (pkg) => {
-		if (pkg.type === 'sourceless')
+		if (pkg.type === 'sourceless' || pkg.type === PackageType_Python)
 			return;
 
 		const folder = 'main';
@@ -504,6 +539,9 @@ export const Phase_PreCompile: BuildPhase = {
 		return fs.existsSync(`${pkg.path}/prebuild.sh`);
 	},
 	action: async (pkg) => {
+		if (pkg.type === PackageType_Python)
+			return;
+
 		return NVM.createCommando(Cli_Basic)
 			.cd(pkg.path)
 			.append(`bash ${pkg.path}/prebuild.sh`).execute();
@@ -518,7 +556,7 @@ export const Phase_Compile: BuildPhase = {
 	action: async (pkg) => {
 		const packages = MemKey_Packages.get();
 
-		if (pkg.type === 'sourceless')
+		if (pkg.type === 'sourceless' || pkg.type === PackageType_Python)
 			return;
 
 		const folder = 'main';
@@ -565,12 +603,17 @@ export const Phase_Compile: BuildPhase = {
 
 			if (runTimePackages) {
 				for (const rtPack of runTimePackages) {
+					if (rtPack.type === PackageType_Python)
+						return;
+
+					// @ts-ignore
 					if (!(rtPack as RuntimePackage_WithOutput).output)
 						continue;
 
 					const pkgOutputFolderAsDependency = `${pkg.output}/.dependencies/${rtPack.name}/`;
 					await NVM.createCommando()
 						.append(`mkdir -p ${pkgOutputFolderAsDependency}`)
+						// @ts-ignore
 						.append(`rsync -a --delete ${(rtPack as RuntimePackage_WithOutput).output}/ ${pkg.output}/.dependencies/${rtPack.name}/`)
 						.execute();
 
@@ -610,13 +653,13 @@ export const Phase_CompileWatch: BuildPhase = {
 
 			const rtPackages = MemKey_Packages.get();
 			const pkg = flatArray(rtPackages.packagesDependency).find(pkg => {
-				return path.startsWith(pkg.path) && pkg.type !== 'sourceless';
+				return path.startsWith(pkg.path) && pkg.type !== 'sourceless' && pkg.type !== PackageType_Python;
 			});
 			if (deleteDist && pkg && 'output' in pkg)
 				await _fs.rmdir(pkg.output);
 
 			const packageIndex = rtPackages.packagesDependency.findIndex(packages => {
-				return packages.some(pkg => path.startsWith(pkg.path) && pkg.type !== 'sourceless');
+				return packages.some(pkg => path.startsWith(pkg.path) && pkg.type !== 'sourceless' && pkg.type !== PackageType_Python);
 			});
 
 			try {
@@ -682,13 +725,15 @@ export const Phase_CompileWatch: BuildPhase = {
 	}
 };
 
+const executorMap: TypedMap<CommandExecutor_FirebaseFunction> = {};
+
 let runningAppsLogs: RunningProcessLogs;
 export const Phase_Launch: BuildPhase = {
 	type: 'package',
 	name: 'launch',
 	terminatingPhase: true,
 	mandatoryPhases: [Phase_ResolveEnv],
-	filter: async (pkg) => !!pkg.name.match(new RegExp(RuntimeParams.launch))?.[0] && (pkg.type === 'firebase-functions-app' || pkg.type === 'firebase-hosting-app'),
+	filter: async (pkg) => !!pkg.name.match(new RegExp(RuntimeParams.launch))?.[0] && (pkg.type === 'firebase-functions-app' || pkg.type === 'firebase-hosting-app' || pkg.type === PackageType_Python),
 	action: async (pkg) => {
 		const projectManager = MemKey_ProjectManager.get();
 		const projectScreen = MemKey_ProjectScreen.get();
@@ -714,22 +759,42 @@ export const Phase_Launch: BuildPhase = {
 		projectScreen.updateOrCreatePackage(pkg.name, 'Launching...');
 
 		if (pkg.type === 'firebase-functions-app') {
-			runningAppsLogs.registerApp(pkg.name, logClient);
+			runningAppsLogs?.registerApp(pkg.name, logClient);
 			const executor = await new CommandExecutor_FirebaseFunction(pkg).execute();
-			runningAppsLogs.addOnTerminateCallback(async () => {
+			executorMap[pkg.name] = executor;
+			runningAppsLogs?.addOnTerminateCallback(async () => {
 				await executor.kill();
-				runningAppsLogs.unregisterApp(pkg.name);
+				runningAppsLogs?.unregisterApp(pkg.name);
 			});
 			return;
 		}
 
 		if (pkg.type === 'firebase-hosting-app') {
-			runningAppsLogs.registerApp(pkg.name, logClient);
+			runningAppsLogs?.registerApp(pkg.name, logClient);
 			const executor = await new CommandExecutor_FirebaseHosting(pkg).execute();
-			runningAppsLogs.addOnTerminateCallback(async () => {
+			runningAppsLogs?.addOnTerminateCallback(async () => {
 				await executor.kill();
-				runningAppsLogs.unregisterApp(pkg.name);
-			})
+				runningAppsLogs?.unregisterApp(pkg.name);
+			});
+		}
+
+		if (pkg.type === PackageType_Python) {
+			runningAppsLogs?.registerApp(pkg.name, logClient);
+			logClient.log(pkg.name, LogLevel.Info, true, ['Awaiting Advisor BE launch']);
+			const advisorExecutor = executorMap['advisor-backend'];
+
+			if (!advisorExecutor)
+				logClient.log(pkg.name, LogLevel.Error, true, ['Advisor executor not registered yet']);
+
+			advisorExecutor?.addOnReadyCallback(async () => {
+				logClient.log(pkg.name, LogLevel.Info, true, ['Advisor BE launched, Starting!']);
+				const executor = new CommandExecutor_Python(pkg);
+				runningAppsLogs?.addOnTerminateCallback(async () => {
+					await executor.kill();
+					runningAppsLogs?.unregisterApp(pkg.name);
+				});
+				await executor.execute();
+			});
 		}
 
 		projectScreen.updateOrCreatePackage(pkg.name, 'Died');
